gh-111165: Move test running code from test.support to libregrtest
Remove no longer used functions run_unittest() and run_doctest() from
the test.support module.
