+++ Misc/NEWS.d/next/Tests/2023-10-21-19-27-36.gh-issue-111165.FU6mUk.rst
@@ -1,0 +1,2 @@
+Remove no longer used functions ``run_unittest()`` and ``run_doctest()``
+from the :mod:`test.support` module.
